+++ test_core_models/tests/test_models_base.py
@@ -54,7 +54,8 @@ def test_historical_protocol_update(historical_model):
     assert created1 == created2 == created3
 
 
-def test_historical_protocol_history(historical_model):
+def test_historical_protocol_history(historical_model, settings):
+    settings.SOFT_DELETE_SAFE_MODE = False
     model, factory = historical_model
     obj1 = factory.create()
     obj2 = factory.create()